+++ CHANGELOG.md
@@ -26,7 +26,7 @@ Thank you to all who have contributed!
 ## [Unreleased](https://TODO.com) - YYYY-MM-DD
 
 ### Added
-- Support aggregation function calls in the `ORDER BY` and `HAVING` clause
+- Support aggregation function calls in the `ORDER BY` clause
 
 ### Changed