+++ CHANGELOG.md
@@ -26,6 +26,23 @@ Thank you to all who have contributed!
 ## [Unreleased](https://TODO.com) - YYYY-MM-DD
 
 ### Added
+
+### Changed
+
+### Deprecated
+
+### Fixed
+
+### Removed
+
+### Security
+
+### Contributors
+Thank you to all who have contributed!
+
+## [1.3.1](https://TODO.com) - 2025-11-06
+
+### Added
 - Support aggregation function calls in the `ORDER BY` and `HAVING` clause
 - Add Support to transcribe `Overlaps` function
 - Add planner support to compare DATE and TIMESTAMP(Z)